+++ src/main/java/com/buuz135/industrial/plugin/jei/category/DissolutionChamberCategory.java
@@ -51,7 +51,6 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import java.awt.*;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class DissolutionChamberCategory implements IRecipeCategory<DissolutionChamberRecipe> {
@@ -89,8 +88,6 @@ public class DissolutionChamberCategory implements IRecipeCategory<DissolutionCh
     @Override
     public void setRecipe(IRecipeLayoutBuilder builder, DissolutionChamberRecipe recipe, IFocusGroup focuses) {
         for (int i = 0; i < recipe.input.size(); i++) {
-            ItemStack stack = Arrays.stream(recipe.input.get(i).getItems()).toList().getFirst();
-            stack.getItem().onCraftedBy(stack, null, null);
             builder.addSlot(RecipeIngredientRole.INPUT, 24 + DissolutionChamberTile.getSlotPos(i).getLeft(), 11 + DissolutionChamberTile.getSlotPos(i).getRight()).addIngredients(recipe.input.get(i));
         }
         if (recipe.inputFluid != null && !recipe.inputFluid.isEmpty()) {
@@ -98,7 +95,6 @@ public class DissolutionChamberCategory implements IRecipeCategory<DissolutionCh
         }
         if (!recipe.output.isEmpty()) {
             ItemStack stack = recipe.output.get();
-            stack.getItem().onCraftedBy(stack, null, null);
             builder.addSlot(RecipeIngredientRole.OUTPUT, 119, 16).addIngredient(VanillaTypes.ITEM_STACK, stack);
         }
         if (recipe.outputFluid != null && !recipe.outputFluid.isEmpty()) {