+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 3.6.21
+
+* Changed how item components are added to items
+* Modified Infinity Backpack recipe easier to make
+* Added ingot tags to pink slime
+
 # Version 3.6.20
 
 * Improve Hydroponic Bed code efficiency